% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdist.R
\name{as_cdist}
\alias{as_cdist}
\alias{as_cdist.cdist}
\alias{as_cdist.cdist_item}
\alias{as_cdist.list}
\title{Coerce to a continuous distribution vector}
\usage{
as_cdist(x, ...)

\method{as_cdist}{cdist}(x, ...)

\method{as_cdist}{cdist_item}(x, ...)

\method{as_cdist}{list}(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Passed to other methods}
}
\value{
A continuous distribution vector
}
\description{
Coerce to a continuous distribution vector
}
