% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdist.R
\name{boxplot.cdist}
\alias{boxplot.cdist}
\title{Boxplot a continuous distribution vector using base graphics}
\usage{
\method{boxplot}{cdist}(x, whisker_quantiles = c(0.01, 0.99),
  box_quantiles = c(0.25, 0.75), mid_quantile = 0.5, ...)
}
\arguments{
\item{x}{A continuous distribution vector}

\item{whisker_quantiles}{Quantiles for the ends of the whiskers}

\item{box_quantiles}{Quantiles for the ends of the box}

\item{mid_quantile}{Quantile for the middle marker}

\item{...}{Passed to \link[graphics]{bxp}}
}
\description{
Boxplot a continuous distribution vector using base graphics
}
