% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_curve.R
\name{calibration_curve}
\alias{calibration_curve}
\title{Create a calibration curve}
\usage{
calibration_curve(.data = NULL, cal_age, measured_age,
  measured_age_error = 0, name = NULL)
}
\arguments{
\item{.data}{An optional data frame}

\item{cal_age}{The age in calibrated years BP (before 1950)}

\item{measured_age}{The measured age (usually 14C years  measurements)}

\item{measured_age_error}{Error on the measured age, plus or minus}

\item{name}{a name for the calibration curve}
}
\value{
A tibble classed as an age_calibration_curve
}
\description{
Create a calibration curve
}
