% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdist.R
\name{cdist}
\alias{cdist}
\title{Create custom probability distributions}
\usage{
cdist(..., dist = "norm", density_function = NULL,
  quantile_function = NULL)
}
\arguments{
\item{...}{Parameters to pass to density and quantile functions}

\item{dist}{Suffix for distribution (e.g., "norm", "t", etc.)}

\item{density_function}{Function like \link[stats]{dnorm}. Overrides dist argument.}

\item{quantile_function}{Function like \link[stats]{qnorm}. Overrides dist argument.}
}
\description{
Create custom probability distributions
}
