% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdist.R
\name{cdist_item}
\alias{cdist_item}
\title{Create custom probability distributions}
\usage{
cdist_item(..., dist = NULL, density_function = NULL,
  quantile_function = NULL, dist_info = list())
}
\arguments{
\item{...}{Parameters to pass to density and quantile functions}

\item{dist}{Suffix for distribution (e.g., "norm", "t", etc.)}

\item{density_function}{Function like \link[stats]{dnorm}. Overrides dist argument.}

\item{quantile_function}{Function like \link[stats]{qnorm}. Overrides dist argument.}

\item{dist_info}{Meta information about the distribution}
}
\description{
Create custom probability distributions
}
