% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdist.R
\name{cdist_item_from_densities}
\alias{cdist_item_from_densities}
\title{Create custom probability distributions from discrete densities}
\usage{
cdist_item_from_densities(.data = NULL, values, densities)
}
\arguments{
\item{.data}{A data frame from which to source values and densities}

\item{values, densities}{Values and relative probability densities}
}
\description{
Create custom probability distributions from discrete densities
}
