% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdist.R
\name{cdist_norm}
\alias{cdist_norm}
\alias{cdist_t}
\title{Construct parameterized distribution vectors}
\usage{
cdist_norm(.data = NULL, mean, sd, names = NULL)

cdist_t(.data = NULL, df, ncp, names = NULL)
}
\arguments{
\item{.data}{An optional data frame from which parameters should be sourced}

\item{mean, sd}{Parameters of the normal distribution}

\item{names}{Names for output objects}

\item{ncp, df}{Parameters of the t distribution}
}
\description{
Construct parameterized distribution vectors
}
