% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdist.R
\name{custom_cdist}
\alias{custom_cdist}
\title{Create custom probability distributions from discrete densities}
\usage{
custom_cdist(.data = NULL, values, densities)
}
\arguments{
\item{.data}{A data frame from which to source values and densities}

\item{values, densities}{Values and relative probability densities}
}
\description{
Create custom probability distributions from discrete densities
}
