% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdist.R
\name{density.dist_item_parameterized}
\alias{density.dist_item_parameterized}
\alias{density.dist_item_custom}
\title{Density functions for continuous distribution items}
\usage{
\method{density}{dist_item_parameterized}(x, values, ...)

\method{density}{dist_item_custom}(x, values, ...)
}
\arguments{
\item{x}{A continuous distribution item}

\item{values}{Values at which densities should be calculated}

\item{...}{Unused}
}
\value{
A numeric vector of densities
}
\description{
Density functions for continuous distribution items
}
