% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdist.R
\name{dist_item_parameterized}
\alias{dist_item_parameterized}
\alias{dist_item_custom}
\title{Construct individual continuous distribution items}
\usage{
dist_item_parameterized(dist, params)

dist_item_custom(values, densities)
}
\arguments{
\item{dist}{The distribution (e.g., norm)}

\item{params}{Parameters of the distribution (e.g. mean, sd)}

\item{values}{Values for custom density distributions}

\item{densities}{Paired density values for values}
}
\value{
A continuous distribution object
}
\description{
Construct individual continuous distribution items
}
