% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdist.R
\name{dist_norm}
\alias{dist_norm}
\alias{dist_t}
\title{Construct parameterized distribution vectors}
\usage{
dist_norm(.data = NULL, mean, sd, names = NULL)

dist_t(.data = NULL, m, s, df, names = NULL)
}
\arguments{
\item{.data}{An optional data frame from which parameters should be sourced}

\item{mean, sd}{Parameters of the normal distribution}

\item{names}{Names for output objects}

\item{m, s, df}{Parameters of the t distribution (location of m, scale of s)}
}
\description{
Construct parameterized distribution vectors
}
