% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_curve.R
\name{in_cal_bp}
\alias{in_cal_bp}
\alias{in_year_ad}
\alias{in_year_ad.default}
\alias{in_cal_bp.default}
\alias{in_cal_bp.age_calibration_curve}
\alias{in_year_ad.age_calibration_curve}
\title{Convert an age calibration curve to a different calibrated age type}
\usage{
in_cal_bp(x)

in_year_ad(x)

\method{in_year_ad}{default}(x)

\method{in_cal_bp}{default}(x)

\method{in_cal_bp}{age_calibration_curve}(x)

\method{in_year_ad}{age_calibration_curve}(x)
}
\arguments{
\item{x}{An existing age calibration curve (\link{calibration_curve}) or numeric vector}
}
\value{
A new age calibration curve
}
\description{
Convert an age calibration curve to a different calibrated age type
}
