% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdist.R
\name{new_cdist}
\alias{new_cdist}
\title{Create a continuous distribution vector}
\usage{
new_cdist(x)
}
\arguments{
\item{x}{A list}
}
\value{
A continuous distribution vector
}
\description{
Create a continuous distribution vector
}
