% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{plot.calibrate_result}
\alias{plot.calibrate_result}
\title{Plot a calibration result}
\usage{
\method{plot}{calibrate_result}(x, ..., max_plot = 9, n_col = 3)
}
\arguments{
\item{x}{An age calibration}

\item{...}{Passed to \link[graphics]{plot}}

\item{max_plot}{Maximum number of items to plot}

\item{n_col}{Number of columns}
}
\description{
Plot a calibration result
}
