% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdist.R
\name{quantile.dist_item_parameterized}
\alias{quantile.dist_item_parameterized}
\alias{quantile.dist_item_custom}
\title{Quantile functions for continuous distribution items}
\usage{
\method{quantile}{dist_item_parameterized}(x, probs = seq(0, 1, by = 0.25),
  names = FALSE, ...)

\method{quantile}{dist_item_custom}(x, probs = seq(0, 1, by = 0.25),
  names = FALSE, ...)
}
\arguments{
\item{x}{The distribution}

\item{probs}{The probabilities}

\item{names}{Whether or not to include names in output}

\item{...}{Unused}
}
\value{
A numeric vector (possibly named) of quantiles
}
\description{
Quantile functions for continuous distribution items
}
