% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_curve.R
\name{read_14c}
\alias{read_14c}
\title{Read a Carbon-14 Calibration Curve}
\usage{
read_14c(path)
}
\arguments{
\item{path}{A path (or URL) to the calibration curve}
}
\value{
A \link{calibration_curve}
}
\description{
These files, typically ending in .14c, are files with a very specific
format suitable for use with age calibration software.
}
