% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{resolve_curve}
\alias{resolve_curve}
\title{Resolve a calibration curve}
\usage{
resolve_curve(curve, null_age_type = "Calibrated BP", env = parent.frame())
}
\arguments{
\item{curve}{A calibration curve, a character vector with an object name in this package
or the calling environment, or NULL for the identity \link{null_calibration_curve}.}

\item{null_age_type}{The input age type for the NULL curve.}

\item{env}{The environment to look for character names of calibration curves}
}
\value{
A calibration curve
}
\description{
Resolve a calibration curve
}
