% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdist.R
\name{summary.cdist}
\alias{summary.cdist}
\alias{as_tibble.cdist}
\alias{as.data.frame.cdist}
\alias{as.character.cdist}
\alias{format.cdist}
\alias{print.cdist}
\title{Summarise, coerce continuous distribution vectors}
\usage{
\method{summary}{cdist}(object, ..., .id = ".name")

\method{as_tibble}{cdist}(x, ...)

\method{as.data.frame}{cdist}(x, ...)

\method{as.character}{cdist}(x, alpha = 0.05, digits = NULL, ...)

\method{format}{cdist}(x, quote = FALSE, alpha = 0.05, digits = 3, ...)

\method{print}{cdist}(x, alpha = 0.05, digits = 3, ...)
}
\arguments{
\item{object, x}{A \link{cdist} vector}

\item{...}{Passed to parent functions}

\item{.id}{column in which names should be placed, if present}

\item{alpha}{level of confience for character representation of objects}

\item{digits}{Number of digits to display}
}
\description{
Summarise, coerce continuous distribution vectors
}
