% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdist.R
\name{summary.cdist_item}
\alias{summary.cdist_item}
\alias{as.character.cdist_item}
\alias{print.cdist_item}
\alias{format.cdist_item}
\title{Summarise, print, characterify distributions}
\usage{
\method{summary}{cdist_item}(object, ...)

\method{as.character}{cdist_item}(x, digits = 3, alpha = 0.05, ...)

\method{print}{cdist_item}(x, ...)

\method{format}{cdist_item}(x, ...)
}
\arguments{
\item{object, x}{A \link{cdist_item} object.}

\item{...}{Passed to parent methods}

\item{digits}{The number of digits to display}

\item{alpha}{The confidence level to which print when coercing to character}
}
\description{
Summarise, print, characterify distributions
}
