% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{translate_distribution}
\alias{translate_distribution}
\title{Translate a distribution using a calibration curve}
\usage{
translate_distribution(age_calibration_curve, dist, eps = 1e-08, n = 512)
}
\arguments{
\item{age_calibration_curve}{An age_calibration_curve (\link{calibration_curve}) or NULL
for the identity curve}

\item{dist}{A \link{cdist} containing source densities}

\item{eps}{Initial densities smaller than this number will not be considered}

\item{n}{The number of equally-spaced x values at which density should be calculated}
}
\description{
Translate a distribution using a calibration curve
}
