% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdist.R
\name{validate_cdist}
\alias{validate_cdist}
\title{Validate a continuous distribution vector}
\usage{
validate_cdist(x)
}
\arguments{
\item{x}{A continuous distribution vector}
}
\value{
The input, invisibly
}
\description{
Validate a continuous distribution vector
}
