% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_curve.R
\name{calibration_curve}
\alias{calibration_curve}
\title{Create a calibration curve}
\usage{
calibration_curve(.data = NULL, cal_age, measured_age,
  measured_age_error = NULL, measured_age_min = NA_real_,
  measured_age_max = NA_real_)
}
\arguments{
\item{.data}{An optional data frame}

\item{cal_age}{The age in calibrated years BP (before 1950)}

\item{measured_age}{The measured age (usually 14C years  measurements)}

\item{measured_age_error, measured_age_min, measured_age_max}{Error on the measured age, plus or minus}
}
\value{
A tibble classed as an age_calibration_curve
}
\description{
Create a calibration curve
}
