% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdist.R
\name{cdist}
\alias{cdist}
\title{Construct a continuous distribution vector}
\usage{
cdist(...)
}
\arguments{
\item{...}{Items created with \link{cdist_item} or \link{cdist_item_from_densities}.}
}
\value{
A list-ish object of type cdist.
}
\description{
Construct a continuous distribution vector
}
