% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_curve.R
\name{print.age_calibration_curve}
\alias{print.age_calibration_curve}
\alias{plot.age_calibration_curve}
\title{Print, plot an age-depth calibration object}
\usage{
\method{print}{age_calibration_curve}(x, ...)

\method{plot}{age_calibration_curve}(x, type = "l", xlab = "Age (cal BP)",
  ylab = "Measured Age", title = attr(x, "curve_name"), ...,
  measured_age_limits = TRUE)
}
\arguments{
\item{x}{An age-depth calibration}

\item{type, xlab, ylab, ...}{Passed to \link[graphics]{plot}}

\item{title}{A title for the plot}

\item{measured_age_limits}{Plot high and low values of the measured age}
}
\value{
The input, invisibly
}
\description{
Print, plot an age-depth calibration object
}
