% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdist.R
\name{summary.cdist}
\alias{summary.cdist}
\alias{as.character.cdist}
\alias{print.cdist}
\alias{format.cdist}
\title{Summarise, print, characterify distributions}
\usage{
\method{summary}{cdist}(object, ...)

\method{as.character}{cdist}(x, digits = 3, alpha = 0.05, ...)

\method{print}{cdist}(x, ...)

\method{format}{cdist}(x, ...)
}
\arguments{
\item{object, x}{A \link{cdist} object.}

\item{...}{Passed to parent methods}

\item{digits}{The number of digits to display}

\item{alpha}{The confidence level to which print when coercing to character}
}
\description{
Summarise, print, characterify distributions
}
