% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdist.R
\name{translate_distribution}
\alias{translate_distribution}
\title{Translate a distribution from y to x along a discrete function}
\usage{
translate_distribution(.data = NULL, x, y, y_sd = 0, dist, eps = 1e-08,
  n = 512)
}
\arguments{
\item{.data}{A data frame from which to source x, y, and y_sd}

\item{x}{A vector of destination values}

\item{y}{A vector of source values (x->y must be unique)}

\item{y_sd}{Optional standard deviation of source value curve}

\item{dist}{A \link{cdist} containing source densities}

\item{eps}{Initial densities smaller than this number will not be considered}

\item{n}{The number of equally-spaced x values at which density should be calculated}
}
\description{
Translate a distribution from y to x along a discrete function
}
